feat(menus): 优化菜单管理功能并添加缓存支持
- 添加菜单数据传输对象的字段验证注解
- 实现菜单缓存配置和缓存预热功能
- 优化菜单服务层的缓存注解和数据处理逻辑
- 重构控制器参数验证和响应处理机制
- 添加分类VO的排序和创建时间字段
- 修复批量分配角色DTO的验证注解配置
- 优化权限分配相关DTO的验证逻辑
- 更新菜单控制器的参数校验和方法实现

+++ src/main/java/com/ryu/blog/service/ViewHistoryService.java
@@ -71,16 +71,16 @@ public interface ViewHistoryService {
     Mono<ViewHistoryStatsVO> getViewHistoryStats();
 
     /**
-     * 获取文章当前浏览量（优先从Redis获取，Redis无数据则从数据库获取）
+     * 获取文章当前浏览量（优先从缓存获取，缓存无数据则从数据库获取）
      * 
      * @param articleId 文章ID
      * @return 浏览量
      */
-    Mono<Integer> getArticleCurrentViews(Long articleId);
+    Mono<Long> getArticleCurrentViews(Long articleId);
 
     /**
-     * 批量同步Redis中的文章浏览量到数据库
-     * 用于定时任务调用，确保数据库中的浏览量与Redis保持同步
+     * 批量同步缓存中的文章浏览量到数据库
+     * 用于定时任务调用，确保数据库中的浏览量与缓存保持同步
      * 
      * @return 同步的文章数量
      */
